pyproject.toml
[tool.poetry]
name = "days-count"
version = "0.1.0"
description = ""
authors = ["Maksim Panin <me@naxrevlis.com>"]

[tool.poetry.dependencies]
python = "^3.10"
rocketry = "^2.3.0"
pika = "^1.3.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
